users_new_app/requirements.txt
black
cryptography
docker
docker-compose
fastapi
python-dotenv
uvicorn
fastapi_jwt_auth
python-jose
passlib
motor
grpcio
python-grpc
requests-async